replaced json with jsonpickle

+++ firebase/__init__.py
@@ -10,6 +10,7 @@ def close_process_pool():
     Clean up function that closes and terminates the process pool
     defined in the ``async`` file.
     """
-    process_pool.close()
-    process_pool.join()
-    process_pool.terminate()
+    if process_pool:
+        process_pool.close()
+        process_pool.join()
+        process_pool.terminate()
